+++ changelog.md
@@ -5,6 +5,7 @@ Features
 ---------
 * Add `--ssh-options` CLI argument to pass extra options with `--ssh-jump`.
 * Make `ssh_jump` a DSN query parameter.
+* Warn on unknown DSN query parameters.
 
 
 Bug Fixes